+++ src/joiner/classes.ts
@@ -1961,10 +1961,10 @@ export class LPointerTargetable<Context extends LogicContext<DPointerTargetable>
     */
     public dependencies(): Dependency[] { return []; }
     protected get_dependencies(context: Context): () => Dependency[] {
-        const data = context.proxyObject;
+        const data = context.data;
         const dependencies: Dependency[] = [];
         const ret = () => {
-            for(let pointedBy of data.pointedBy) {
+            for (let pointedBy of data.pointedBy) {
                 const raw = pointedBy.source.split('.');
                 let root = raw[0];
                 const obj = raw[1] || '';
@@ -1989,15 +1989,16 @@ export class LPointerTargetable<Context extends LogicContext<DPointerTargetable>
     public delete(): void {}
     protected get_delete(context: Context): () => void {
         const data: LPointerTargetable & GObject = context.proxyObject;
-        const dependencies = data.dependencies();
+        const dependencies = this.get_dependencies(context)();
 
         const ret = () => {
             if (context.data.id.indexOf('Pointer_View') !== -1 ) return; // cannot delete default views/viewpoints
+            if (context.data.__readonly) return;
             for (let child of data.children) {
                 child.delete();
                 // todo: if a m1-dvalue which conforms to a m2-reference with "containment" is deleted, need to delete also target.
                 // maybe better to do through override?
-                child.node?.delete();
+                // child.node?.delete();
             }
 
             for (let dependency of dependencies) {
@@ -2006,7 +2007,7 @@ export class LPointerTargetable<Context extends LogicContext<DPointerTargetable>
                 const field = dependency.field;
                 const op = dependency.op;
                 const val = (op === '-=') ? data.id : '';
-                if((root === 'idlookup') && obj && field) {
+                if ((root === 'idlookup') && obj && field) {
                     console.log('Delete', `SetFieldAction.new('${obj}', '${field}', '${val}', '${op}');`);
                     SetFieldAction.new(obj, field, val, op, false);
                 } else {
@@ -2016,7 +2017,7 @@ export class LPointerTargetable<Context extends LogicContext<DPointerTargetable>
             }
             if (data.nodes) data.nodes.map((node: any) => node.delete());
             SetRootFieldAction.new('ELEMENT_DELETED', data.id, '+=', false); // here no need to IsPointer because it only affects Transient stuff
-            SetRootFieldAction.new('idlookup', data.id, '-=', false); // damiano: shouldn't be isPointer = true?
+            //SetRootFieldAction.new('idlookup', data.id, '-=', false);
             DeleteElementAction.new(data.id);
         };
         return () => TRANSACTION(ret);
